Do something to fix single rgb led mode
LEDC seemed to misbehave previously. Symptoms were:
Red led would become stuck on, others off. Best trigger
was to send 255,0,0 to get the led to red,
then 0,255,0 to have the led turn green, and
finally send 255,255,255 in an attempt to get to white,
but it actually resulted in red.

Some source somewhere was saying that consecutive
LEDC_CHANNELs could be problematic. 0, 2 and 4 didn't
seem to work, but randomly switching to 1, 3 and 5
seems to get the job done.